refreshing token

+++ resources/user.py
@@ -1,9 +1,9 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from passlib.hash import pbkdf2_sha256
-from flask_jwt_extended import create_access_token
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt, get_jwt_identity, create_refresh_token
 from db import db
+from blocklist import BLOCKLIST
 from models import UserModel
 from schema import UserSchema
 
@@ -28,10 +28,27 @@ class UserLogin(MethodView):
     def post(self,user_data):
         user=UserModel.query.filter(UserModel.username==user_data['username']).first()
         if user and pbkdf2_sha256.verify(user_data['password'],user.password):
-            access_token=create_access_token(identity=user.id)
-            return {"access_token": access_token}
+            access_token=create_access_token(identity=user.id, fresh=True)
+            refresh_token=create_refresh_token(identity=user.id)
+            return {"access_token": access_token, "refresh_token": refresh_token}
         abort(401, message="Invalid credentials.")
 
+@blp.route("/refresh")
+class Refresh(MethodView):
+     @jwt_required(refresh=True)
+     def post(self):
+        current_user = get_jwt_identity()
+        new_token = create_access_token(identity=current_user, fresh=False)
+        return {"access_token": new_token}
+
+@blp.route("/logout")
+class UserLogout(MethodView):
+    @jwt_required()
+    def post(self):
+        jti = get_jwt()["jti"]
+        BLOCKLIST.add(jti)
+        return {"message":"Successully logged out."}
+        
 
 
 @blp.route("/user/<int:user_id>")
@@ -43,8 +60,11 @@ class User(MethodView):
         user=UserModel.query.get_or_404(user_id)
         return user
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def delete(self,user_id):
+        jwt=get_jwt
+        if not jwt.__get__("is_admin"):
+            abort(401,message="Admin Privilege required.")
         user=UserModel.query.get_or_404(user_id)
         db.session.delete(user)
         db.session.commit()
